requirements.txt
asgiref==3.8.1
Django==4.2.16
pip==24.3.1
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
setuptools==58.1.0
sqlparse==0.5.2
typing_extensions==4.12.2

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.10.1
django-cors-headers==4.6.0

django-filter==24.3